Refactor downloadScheduler.py, musicapp.html, models.py, and routes.py
