+++ webapp/templates/musicapp.html
@@ -59,7 +59,7 @@
                 <ul class="list-group list-group-flush">
                     <li class="list-group-item">{{music.url}}</li>
                     
-                            <script>
+                            <!-- <script>
                                 $(document).ready(function(){
                                     var musicId = "{{ music.id }}";
                                     var intervalStatusElement = $("#intervalStatus" + musicId);
@@ -100,7 +100,7 @@
                                     });
         
                                 });
-                                </script>
+                                </script> -->
 
 
                                 <div class="progress">
